feat(devtools): align DS1 and DS2 demos to DS3 standard
Add WCAG contrast pass/fail indicators, color swatch cards, type scale
tables, and extracted session demos to bring DS1 and DS2 to structural
parity with DS3. Extract inline data to named constants and replace
inline styles with CSS class-driven markup throughout.
